#102: Refactoring, fixing tests.
