+++ bin/run_qc.py
@@ -4,10 +4,9 @@ import os
 import argparse
 import logging
 
-
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from analysis_driver.quality_control import SamtoolsDepth
-from analysis_driver.dataset_scanner import NoCommunicationDataset
+from analysis_driver.dataset import NoCommunicationDataset
 from analysis_driver.quality_control.gender_validation import GenderValidation
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from analysis_driver.quality_control.genotype_validation import GenotypeValidation
@@ -34,7 +33,7 @@ def _parse_args():
 
     geno_val_parser = subparsers.add_parser('genotype_validation')
     geno_val_parser.add_argument('--project_id', required=True)
-    geno_val_parser.add_argument('--sample_id', required = True)
+    geno_val_parser.add_argument('--sample_id', required=True)
     geno_val_parser.add_argument('--check_neighbour', action='store_true', default=False)
     geno_val_parser.add_argument('--check_project', action='store_true', default=False)
     geno_val_parser.add_argument('--check_samples', nargs='*')
@@ -150,6 +149,7 @@ def run_species_contamination_check(args):
     fastqscreen_result = get_fastqscreen_results(expected_output_files, args.sample_id)
     print(fastqscreen_result)
 
+
 def run_sample_contamination_check(args):
     if args.work_dir:
         work_dir = args.work_dir
@@ -173,6 +173,7 @@ def run_gender_validation(args):
     s = GenderValidation(dataset, work_dir, args.vcf_file)
     s.start()
     return s.join()
+
 
 def median_coverage(args):
     work_dir = os.path.join(cfg['jobs_dir'], args.sample_id)
