+++ tests/test_quality_control/qc_tester.py
@@ -1,15 +1,20 @@
 from unittest.mock import patch
-from analysis_driver.dataset_scanner import SampleDataset, MostRecentProc
+from analysis_driver.dataset import SampleDataset
 from tests import TestAnalysisDriver
 
 
 sample_read_data = 'tests.test_quality_control.qc_tester.SampleDataset._read_data'
 
 
+class NoCommunicationSample(SampleDataset):
+    def _read_data(self):
+        pass
+
+
 class QCTester(TestAnalysisDriver):
     def setUp(self):
         with patch(sample_read_data):
-            self.dataset = SampleDataset(
+            self.dataset = NoCommunicationSample(
                 'test_sample',
                 most_recent_proc={
                     'dataset_type': 'sample',
